% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{center_and_scale}
\alias{center_and_scale}
\title{Function for centering and scaling the transformed metric}
\usage{
center_and_scale(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Object of arbitrary class.
}
\description{
Function for centering and scaling the transformed metric
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
