% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{first_deriv}
\alias{first_deriv}
\title{Take a first derivative, with respect to a variable}
\usage{
first_deriv(dat, var, stdevs)
}
\arguments{
\item{dat}{matrix}

\item{var}{character}

\item{stddevs}{numeric vector}
}
\value{
Object of arbitrary class.
}
\description{
Take a first derivative, with respect to a variable
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
