% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{fixna}
\alias{fixna}
\title{Fix NAs}
\usage{
fixna(x)
}
\arguments{
\item{x}{numeric}
}
\value{
Object of arbitrary class.
}
\description{
Fix NAs
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
