% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{get_group}
\alias{get_group}
\title{Assign a knot group (for quantiles q) to a vector x}
\usage{
get_group(x, q, rev = F)
}
\arguments{
\item{x}{numeric vector}

\item{q}{numeric vector}

\item{rev}{logical}
}
\value{
Object of arbitrary class.
}
\description{
Assign a knot group (for quantiles q) to a vector x
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
