% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{interact2_deriv}
\alias{interact2_deriv}
\title{Take a second derivative for only interactions}
\usage{
interact2_deriv(dat, var, stdevs)
}
\arguments{
\item{dat}{matrix}

\item{var}{character}

\item{stddevs}{numeric vector}
}
\value{
Object of arbitrary class.
}
\description{
Take a second derivative for only interactions
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
