% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{knot_expand2}
\alias{knot_expand2}
\title{Function useful for turning cubic expansions into knot expansions}
\usage{
knot_expand2(mat, Q)
}
\arguments{
\item{mat}{matrix}

\item{Q}{matrix}
}
\value{
Object of arbitrary class.
}
\description{
Function useful for turning cubic expansions into knot expansions
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
