% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{my_rinvgamma}
\alias{my_rinvgamma}
\title{Generate inverse-gamma variables, with error handling}
\usage{
my_rinvgamma(n, a, b, s)
}
\arguments{
\item{n}{numeric}

\item{a}{numeric}

\item{b}{numeric}

\item{s}{numeric}
}
\value{
Object of arbitrary class.
}
\description{
Generate inverse-gamma variables, with error handling
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
