% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{sample_between_intervals}
\alias{sample_between_intervals}
\title{Function for performing latin-grid sampling}
\usage{
sample_between_intervals(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Object of arbitrary class.
}
\description{
Function for performing latin-grid sampling
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
