% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{second_deriv}
\alias{second_deriv}
\title{Compute second derivatives, no interactions}
\usage{
second_deriv(dat, var, stdevs)
}
\arguments{
\item{dat}{matrix}

\item{var}{character}

\item{stddevs}{numeric vector}
}
\value{
Object of arbitrary class.
}
\description{
Compute second derivatives, no interactions
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
