% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{solvetry}
\alias{solvetry}
\title{First attempting a normal inversion, before resorting to the M.P.G.I.}
\usage{
solvetry(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{matrix}

\item{tol}{numeric}
}
\value{
Object of arbitrary class.
}
\description{
First attempting a normal inversion, before resorting to the M.P.G.I.
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
