% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{transf}
\alias{transf}
\title{Variance stabilizing/symmetry-inducing transformations}
\usage{
transf(x, type, pwr_lmbda = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{type}{character}

\item{pwr_lmbda}{optional numeric}
}
\value{
Object of arbitrary class.
}
\description{
Variance stabilizing/symmetry-inducing transformations
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
