% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{un_transf}
\alias{un_transf}
\title{variance stablizing/symmetry-inducing back-transformations}
\usage{
un_transf(x, type, pwr_lmbda = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{type}{character}

\item{pwr_lmbda}{optional numeric}
}
\value{
Object of arbitrary class.
}
\description{
variance stablizing/symmetry-inducing back-transformations
}
\details{
for internal use with set_optim_thompson function
}
\examples{
\donttest{
  # No Examples Shown
}
}
